requirements.txt
flask==2.0.1
python-dotenv==0.19.0
google-generativeai>=0.3.1
fpdf>=1.7.2
numpy>=1.20.0
Pillow>=8.2.0
werkzeug>=2.0.0